WIP for version 4 - add tools as assets for downloading

+++ gettools.py
@@ -13,6 +13,8 @@ TOOL_PATH = os.path.join(ROOT, 'tools')
 LOCAL_ZIP = os.path.join(TOOL_PATH, 'assets.zip')
 LOCAL_JSON = os.path.join(TOOL_PATH, 'assets.json')
 TEMP_JSON = os.path.join(TOOL_PATH, 'temp.json')
+TEMP_ZIP = os.path.join(TOOL_PATH, 'temp.zip')
+
 ASSETS_JSON  = 'https://raw.githubusercontent.com/DrDonk/patchersupportpkg/main/assets/assets.json'
 ASSETS_ZIP  = 'https://raw.githubusercontent.com/DrDonk/patchersupportpkg/main/assets/assets.zip'
 
@@ -29,20 +31,37 @@ def main():
 
     # Get the assets.json file
     urlretrieve(ASSETS_JSON, TEMP_JSON)
+    # Download the assets.zip file
+    urlretrieve(ASSETS_ZIP, TEMP_ZIP)
+
+    # Get the data for current and downloaded zip
     with open(TEMP_JSON) as f:
         new_data = json.load(f)
     with open(LOCAL_JSON) as f:
         old_data = json.load(f)
 
+    # Check that we have a new build
     if old_data['build'] != new_data['build']:
-        # Download the assets.zip file
-        urlretrieve(ASSETS_ZIP, LOCAL_ZIP)
-        with open(LOCAL_ZIP, 'rb') as f:
+        # Check we downloaded OK by using SHA256 checksum
+        with open(TEMP_ZIP, 'rb') as f:
             data = f.read()  # read entire file as bytes
-            readable_hash = hashlib.sha256(data).hexdigest();
-            print(readable_hash)
-        os.remove(LOCAL_JSON)
-        os.rename(TEMP_JSON, LOCAL_JSON)
+            new_hash = hashlib.sha256(data).hexdigest()
+
+        # Replace json and zip file if hash is OK
+        if new_hash == new_data['sha256']:
+            print(f'Successfully downloaded build {new_data["build"]}')
+            if os.path.exists(LOCAL_JSON):
+                os.remove(LOCAL_JSON)
+            os.rename(TEMP_JSON, LOCAL_JSON)
+            if os.path.exists(LOCAL_ZIP):
+                os.remove(LOCAL_ZIP)
+            os.rename(TEMP_ZIP, LOCAL_ZIP)
+        else:
+            print(f'Download of build {new_data["build"]} failed checksoums do not match:')
+            print(f'Expected: {new_data["sha256"]}')
+            print(f'Actual:   {new_hash}')
+            os.remove(TEMP_JSON)
+            os.remove(TEMP_ZIP)
 
     return
